All the changes needed to port/platform/graphics needed to build with ToT WebKit from webkit.org. All of these files will eventually go upstream (after a little cleanup).
Most noteable here is that we are now using WebKit-style
cast-operators to convert between WebCore types (IntRect) and
Skia types (like SkIRect).  This means that SkiaUtils.h will
eventually be completely removed. :)

Note that we can already transparently convert Colors
between SkColor and WebCore::Color because WebCore::RGBA32 is
the same type and layout as SkColor! :)  I've added a compile-time
ASSERT to make sure this is always true.