Get GKURL_unittest.cpp compiling.
(This is a first pass, there is still some work to be done. Notably, need to add implementation for protocolIs() to GKURL.cpp before it can link)


I apologize in advance for the ugliness of this change.

The problems it addresses are:

(1) The new version of KURL.h/KURL.cpp declare/define a number of new global functions.
These globals cause redefinitions when we include both KURL.cpp and GKURL.cpp from GKURL_unittest.cpp
The workaround is to move the globals into the KURL:: namespace to avoid the conflict.

(2) DeprecatedString is no more (w00t), so url.deprecatedString().ascii() doesn't fly anymore.
Changed such references to url.string().utf8().data().

Note: I am not 100% certain if this is going to place nice with temporaries
(potentially EXPECT_STREQ expands to something crazy, and the the temp CString returned by utf8() is destroyed before the data pointer is read?).
Once the merge branch is further along I can revisit and check whether this paradigm is safe.

Note: the GKURL version could instead use "url.utf8String().data()" and avoid the temporary CString, but I opted for consistency with the KURL version.