Fix proxy code. The pac string was corrupted when a pac file was specified.
This is because AsciiToWide returns a temporary object, and
we use .c_str, but as soon as the line is executed, the
object is destroyed and what was pointed by c_str is also
deleted.

Use a local variable instead.

Thanks for griffinz@gmail.com for pointing the issue.

Review URL: http://codereview.chromium.org/1838