This fixes the following bugs:- 1. http://code.google.com/p/chromium/issues/detail?id=1473 2. http://code.google.com/p/chromium/issues/detail?id=1533
Our plugin implementation currently sends the containing frame URL
as the default referer in HTTP requests initiated by plugins. Based
on a discussion on the Plugin futures mailing list, an agreement
has been reached to mimic IE behavior and send in the plugin source
URL as the default referer in plugin requests. If the plugin is instantiated
with an empty SRC, then use the containing frame URL as the referer.

Bug=1473,1533
R=jam

Review URL: http://codereview.chromium.org/3180