Badness all around. This change will have to be reverted...much later.
There is some weird off by one error somewhere regarding the line
number given on CONSOLE MESSAGE: line xxx, where xxx is off by one.
We don't know where and we aren't going to hold up the merge for 
this; filing a todo bug, but this needs to be fixed.

TBR=dglazkov,tony