Attempt 2 at landing this patch. It broke the Mac/linux builds. Fix for now is to add the offending code in np_v8object.cpp in a ifdef OS_WIN
This fixes http://code.google.com/p/chromium/issues/detail?id=2472, which
is an issue with popups displayed by the flash plugin in response to a user
click, getting blocked.

The plugin invokes NPN_Evaluate to execute the javascript. It also invokes
the NPN_PushPopupEnabledState/NPN_PopupEnabledState API's to set the popup
enabled stack for the duration of the call. 

The fix is to add the plumbing in NPN_Evaluate to pass in a flag indicating
whether popups are allowed for the duration of the call.

Bug=2472
R=jam

Review URL: http://codereview.chromium.org/6379