Support zero size chunks in the current implementation.
This allows the SafeBrowsing servers to send us Add and Sub
chunks with no content so that when we report our update
status, the request size is decreased.

Our update status request contains a list of all chunks that
we have received from the service, and this can get fragmented
over time. For example, the part of the request containing our
phishing chunks might look like this:

  goog-phish-shavar:a:1,3,5,7,9,11,13,15,17

By sending zero size chunks for the chunk numbers missing from
the above example, the report will now look like:

  goog-phish-shavar:a:1-17

Given the large number of chunks and the rate of chunk expiry,
this change will reduce the client request size.

BUG= http://code.google.com/p/chromium/issues/detail?id=3262
Review URL: http://codereview.chromium.org/6369