Merge r3117, r3176, and r3178 to the release branch:
r3117: Changing the About box according to mockups. Basically, we now want to 
have a paragraph of text with embedded links (to the Chromium project, credits
and terms of service).

r3176: Allow the translators to chose which link appears first in the About 
box.

r3178: Forgot a .h file.

TBR= finnur@google.com

Review URL: http://codereview.chromium.org/7072