Merge r3295 to the release branch:
The Schannel considers some cipher suites (e.g., the
AES cipher suites) to be TLS only, and won't allow
them to be used in SSL 3.0.  If an SSL 3.0 server
picks such a cipher suite, we need to handle it as
if it were TLS intolerant by turning off TLS and
retrying.

BUG=3203,3001
TBR= wtc@google.com

Review URL: http://codereview.chromium.org/7277