Merge r3528 to the release branch:
Fix bug where a shelf or infobar would not be closed correctly when closing a tab. 

The issue was because ShelfVisibilityChangedImpl is called twice, once with a null current_tab for the closing tab (via TabClosingAt), and once on selection of the new tab (via TabSelectedAt) - the first call nulls the shelves without doing a Layout, the nulled shelves then don't show up in the second call, so Layout is never called.

http://crbug.com/2069

BUG=2069
TBR= glen@google.com
Review URL: http://codereview.chromium.org/7487