Merge r3457, r3458, r3459, and r3460 to the release branch.
We should also pick up r3466 and r4344. Those are queued for merging later.

r3457: This fixes http://code.google.com/p/chromium/issues/detail?id=146&,
which was an issue with navigation attempts initiated by Flash not
working correctly in Chrome.

The plugin would initiate a navigation in response to a user click. This
would eventually result in a call to a script on the page, which would
query the history in a timer, and would end up resetting the selected
URL to the original URL as the history length would not be updated as
yet. The reason being the following:-

1. A Frame navigation is initiated by the renderer in response to the
user click.
2. This is sent as an async message to the browser UI thread.
3. When the navigation in the browser completes we update the history
info in render view asynchronously.

The fix is to attempt to update the history count when we receive 
notifications from webkit. These include notifications for items being 
added to the history, navigations based on history (like 
history.back/forward/goto, etc).

These counts continue to be updated as before from the browser as well.

Added UI tests to test these cases.

Bug=146

r3458, r3459, r3460: Fix up build and test issues.

TBR= ananta@chromium.org
Review URL: http://codereview.chromium.org/9208