Merge r3982 to the release branch:
Added the class AutofillForm.  

AutofillForm contains a list of AutofillForm::Elements each of which holds the name and value of a text input field in a form.  An AutofillForm gets created by a call to a factory function when the user submits a form, and the factory function iterates through the DOM extracting the names and values (current contents) of each text field.

TBR=petersont@google.com

Review URL: http://codereview.chromium.org/9084