Merge r4094 to the release branch:
NOTE: Adding
A  +   chrome\test\data\npapi\execute_script_delete_in_mouse_move.html

For some reason, this merge-adds don't get uploaded to the server.
------------
Proposed fix for http://b/issue?id=1362948, which is a crash in the
rendererwhen we invoke the setCursor call on the parent view in
WebPluginImpl::handleEvent.
This crash occurs because the plugin is deleted in the context of a
mouse down event. This could occur by invoking a javascript function
via NPN_Evaluate. On return from the HandleEvent sync call we
attempt to retreive the parent frame, which returns NULL and hence
the crash.

The fix is to retreive the parent frameview at the start of the
WebPluginImpl::handleMouseEvent function and use it whereever
needed.

Added a unit test which deletes the plugin instance in a mousemove
event

Bug=1362948
TBR= ananta@chromium.org
Review URL: http://codereview.chromium.org/9307