Merge r4249 to the release branch:
When acting as a drop target, the webview keeps a cache of
the drop data so we don't have to send it over the IPC for
each drag move.  However, if a page load happens during
the drop, the webview changes and we lose the cached
data.

To handle this, we keep track of which RVH we're dragging
over and if it changes, we re-send the drop data

TBR=tc@google.com

Review URL: http://codereview.chromium.org/8412