Merge r4517 to the releae branch:
This fixes bug 3907, which was a crash in the browser in the IPC 
handler codepath which handles scrolling.

The crash occured because of a failure to find the backing store 
for the RenderWidgetHost instance. We would end up dereferencing a 
NULL backing store pointer and crash.

The fix for this is to check for a NULL backing store and return. 

Bug=3907
TBR= ananta@chromium.org
Review URL: http://codereview.chromium.org/9176