Merge SafeBrowsing fixes to the release branch:
r2814:
A partially working implementation of a faster safe browsing implementation.

r3031:
Fixes to the experimental bloom filter storage code. Update the unittest to handle the  bloom filter storage implementation and fix a test scenario.

r3131:
Support zero size chunks in the current implementation.

This allows the SafeBrowsing servers to send us Add and Sub chunks with no content so that when we report our update status, the request size is decreased.

Our update status request contains a list of all chunks that we have received from the service, and this can get fragmented over time. For example, the part of the request containing our phishing chunks might look like this:

  goog-phish-shavar:a:1,3,5,7,9,11,13,15,17

By sending zero size chunks for the chunk numbers missing from the above example, the report will now look like:

  goog-phish-shavar:a:1-17

Given the large number of chunks and the rate of chunk expiry, this change will reduce the client request size.

r3148:
Provide an option to turn on the new SafeBrowsing storage system via a command line flag ("--new-safe-browsing").

This is to ease testing and is not yet ready for production use.

r3165:
Fix a couple of UMRs in the unit tests.

r3314:
Minor code clean up.

r3717:
Change the detection of the list name from parsing a URL to storing it from the update in order to avoid breaking due to changes in the format.

r3959:
Changes to allow running the new SafeBrowsing storage system,contained in SafeBrowsingDatabaseBloom, via a command line flag(--new-safe-browsing).

r4032:
Clean up the chunk processing pipeline. We are no longer asynchronous, so we can greatly simplify the code.

r4087:
Give the new implementation a unique name for the database and filter storage so that it can run side by side with the old implementation without data loss.

r4117:
Fix a minor memory leak.


TBR=paulg@google.com

Review URL: http://codereview.chromium.org/9606