Merge r4842 to the release branch:
This fixes bug http://crbug.com/4076, which was a hang while 
loading certain PDF files. This was a regression caused by 
support for NPN_RequestRead (PDF Fast Webview). We had 
incorrectly assumed that the Content Type showing up for 
partial HTTP Responses would always end with the boundary. A 
content type can show up like multipart/byteranges; 
boundary=--bound--; charSet=utf8. As a result we would look 
for the wrong boundary in the actual data resulting in a hang.

The parsing code now accounts for this. Added a unit test to 
test this case.

Bug=4076
TBR= ananta@chromium.org

Review URL: http://codereview.chromium.org/9445