Merge r4912 to the release branch:
Add a function for eliding filenames to a set pixel width while 
preserving the display of the extension. This touches a lot of 
files because it renames url_elider to text_elider (as it already
contained a bunch of text ellision functions), and so requires 
header file reference changes. 

BUG=1208
TBR= glen@chromium.org
Review URL: http://codereview.chromium.org/9681