Merge r5724 to the release branch:
Fixes crash when showing modal dialog from context menu. With this
sequence ContainerWin remains on the stack while the modal dialog is
showing. Prior to this fix we would delete the ContainerWin, so that
when the modal dialog closed control would return to a deleted
object. The fix is to effectively delay deleting of the ConatinerWins
used by menus.

BUG=4580
TEST=This is the sequence that triggered the crash: open a folder on
  the bookmark bar, right child on a child folder that has more than 15
  descendant URLs, click open all, a dialog should appear asking if you
  really want to open all, click cancel and make sure it doesn't crash.
  This is a subtle change to bookmark menus though, would be good to
  thoroughly exercise the menus in as many permutations as you can
  think of.
TBR=sky
Review URL: http://codereview.chromium.org/11327