Merge r5855 to the release branch.
Reason: avoid a potential browser crash

This ports only the change to session_service to check whether a browser
has tabs and a window before we build commands from a browser that is 
about to be removed.

The change in browser.cc does not apply to the release branch.

BUG=4666
TBR=sky
Review URL: http://codereview.chromium.org/11383