Attempted fix for some wheel/touchpad scrolling errors.
The nTrackPos member of SCROLLINFO is not actually the scroll position, and can't be set by a SetScrollInfo() call.  Thus the SetScrollInfo() calls were actually setting the scroll position of the fake scroll bars to the upper left corner.

This change sets the scroll position to the center of each scrollbar, in the hopes that some mouse drivers were querying the scroll position or thumb position and not sending scroll left/up messages when they found the page was already scrolled all the way left/up.

This change may not help, but it won't hurt, and the old code was definitely wrong.
Review URL: http://codereview.chromium.org/12415