Update to latest trunk version of V8.
This includes Evan's change to use char instead of wchar_t for counter
names.

Also, because of changes to the regexp parser, we need to rebaseline a
couple of tests mainly because of error message changes.  Most of the
tests were already rebaselined because of differences in error
messages.
Review URL: http://codereview.chromium.org/12902