Handle destruction of test shell properly.
When test shell's destructor runs, it tries to delete its child objects.
But when the GTK widgets are destroyed, they also try to delete the associated
objects.  We don't want both to run.

Also, there were a couple of incorrect uses of the GTK API (e.g. destroy-event
instead of destroy) that were obfuscating the problem.

This fixes a crash on shutdown, so I'm restoring the LoadURL() call in
TestShell's destructor.

Review URL: http://codereview.chromium.org/13214