Correctly construct platform keyboard event in event sending controller.
Same as issue 12981 with but now with change to test_shell.vsprops. This change makes it actually compile on windows! It also makes it clear that test shell is looking in webkit directories for include files. Apparently this was already the case on linux and mac, but it is new for windows.

Review URL: http://codereview.chromium.org/13678