Merge r6705 to the release branch.
"The m_context member of the V8 proxy should be empty if any part of
the initialization fails.  In particular, if the DOM constructor cache
could not be allocated, initialization of the context fails and we
dispose and clear the m_context member to signal that the context
cannot be used."

NOTE: v8_proxy.cpp did not merge cleanly, so I ported the changes by hand.
There's some risk here. I think it'd be simplest if I pointed it out in 
codereview.

TBR=lager

Review URL: http://codereview.chromium.org/14456