Rebaseline numeric-conversions per ager's advice.
The expected files expect a FAIL line because I can't change the upstream test. Copying the test to the platform/directories doesn't help because the test is actually in a resources/xxx.js file and there seem to be problems with paths in that case.

Review URL: http://codereview.chromium.org/15201