Re-baselining a few SVG batik tests to fix 5 layout tests. Also added a few expected text baselines.
The changes are mainly due to subtle differences in
the way we render text, missing Z's from text diff 
and in some cases I rebaselined because the actual 
output is a lot more like the mac output than our 
expected output and given that when we baselined SVG 
we just put a stake in the ground at some point
so there are errors in some baselines.

In some cases I just added the missing Z's to make 
text baseline correct (although it still fails 
image diff due to other issues).

Review URL: http://codereview.chromium.org/18016