This change list fix issues 6093 [settings] bad directionality for gears button (http://crbug.com/6093).
The directionality of gears button is already fixed.
This CL fixes the wrong location of ":".

It adds RLM at the end of string "Google Gears:" for Hebrew and Arabic Chrome,
so that the ":" in "Google Gears:" in Hebrew is displayed leftmost instead of
rightmost.


Test:
1) open chrome, Hebrew UI
2) option -> under the hood -> google gears

Without the fix, the name appears as "Google Gears:".
With the fix, the name appears as ":Google Gears".


Review URL: http://codereview.chromium.org/18279