Improve SDCH stability stats and error recovery
This CL provides an exponential back-off (within
a run of chromium) for recovery to a decoding
error, rather than an full disable of SDCH when
a (minor) error appears.  This will help to
recover from the common (implementation)
problem where a user has an SDCH encoded
page as a startup tab, but doesn't still
have the requisite dictionary from the 
previous run.  The exponential backoff is
not "time based," but instead counts down
the number of times a tentatively blacklisted
domain is accessed.  

I now detect the use of the browser cache in an
SDCH decode, and use that to better identify if
external proxies have corrupted the process,
vs internal use of cache (such as at startup)
has caused a "dictionary not found" error
to emerge.


r=openvcdiff,huanr
Review URL: http://codereview.chromium.org/18355