Fix a crash when dns prefetching was being shut down without being started.
This happened when BrowserMain exited before the point when we initialized dns prefetching, but after BrowserProcessImpl was created. We do shutdown of dns prefetching in ~BrowserProcessImpl.

Example way to trigger the crash: open a saved HTML page while Chrome is already running.

Change ShutdownDnsPrefetch to EnsureDnsPrefetchingShutdown which does nothing
if dns prefetching  has not been initialized. This is the simplest way
to prevent the crash and we don't want to complicate the startup sequence even more.

BUG=7900

Review URL: http://codereview.chromium.org/27048