alternateFamilyName customization for Windows was landed in Webkit (r40636, https://bugs.webkit.org/show_bug.cgi?id=20531). Instead of adding a static member function to FontCache class to be implemented in each platform, |alternateFamilyName| in FontCache.cpp was changed with #ifdef's for Windows.
This CL make our copy of FontCache.h identical to that of Webkit trunk's. FontCache.cpp was merged in the latest merge. 

It also removes the implementation of alternateFamilyName in FontCache{ChromiumWin,Linux}.cpp. This part of the
patch was uploaded to webkit bug 20531 (https://bugs.webkit.org/attachment.cgi?id=27486 )

FontCacheChromiumWin.cpp also includes a patch for off-by-1 error (webkit bug : http://bugs.webkit.org/show_bug.cgi?id=23803). This was not intended, but the fix is trivial and I'll land it together once the webkit trunk gets it. 


BUG=7549



Review URL: http://codereview.chromium.org/21174