Fix startup_test failures on Windows buildbot.
The problem seems to be an error when overwriting file which is in use
(that's how EvictFileFromSystemCache works on Windows). So I used
a standard workaround for such Windowsic problems and created a wrapper
which retries 10 times. Similar code was there before my porting patch.

TBR=agl

Review URL: http://codereview.chromium.org/40160