Enable interstitial pages for SSL certificate errors on Linux.
There are still problems with this patch, as acknowledging the interstitial page often fails to load a new page. Instead, it just displays a blank gray page. This appears to be somewhat timing related.

Nonetheless, this changelist is strictly an improvement over what we have right now. So, I would still like to commit it.

Ideally, I'd like to remove the special-case Windows code. But I don't have a Windows machine to check whether the new code is actually a full superset of the old Windows code.
Review URL: http://codereview.chromium.org/40095