Fix about box truncation (and truncation of probably every other box on glass frames).
Our windows don't have WS_CAPTION. It's remarkable to me that we've made it this far without noticing this glaring oversight. I guess I fixed a bug with my frame changes earlier in the week that resulted in AdjustWindowRectEx getting called correctly for the first time, and it was not providing correct offsets because of the absence of this flag.

http://crbug.com/8445

Review URL: http://codereview.chromium.org/39300