Initial WebDevToolsAgent implementation contains two agent objects: Dom agent and Net agent.
Dom agent provides API for querying for DOM nodes and receiving notifications
on
Dom updates. It has some logic in and this logic is covered with the unit
tests.

Net agent pushes an initial set of request/response-related events to the
client. It is to be filled with more data later on. It currently caches loaders
for all the requests which is Ok for the case when this agent is turned ON (at
least for now).

Note that this code is not yet wired to the dev tools agent (this is by design).
The plan is to start enrolling the dev tools agent glue that connects these
sub-agents with the IPC transport once this CL is in.

Original CL: http://codereview.chromium.org/41008/show

Review URL: http://codereview.chromium.org/43128