Use anonymous shared memory for VisitedLinks.
We already pass a handle to the shared memory via IPC but, when we
create it, we create a named segment. These segments are not garbage
collected on POSIX.

Since nobody actually uses the name of the segment, use an anonymous
segment instead and remove the functions.

Review URL: http://codereview.chromium.org/48087