Fixing chrome/fast/dom/java-applet-calls.html
I don't understand how the test could have worked because
if you don't specify a CODEBASE for the Java Applet the
default is the url. In the Java console I was seeing it
try to load the class by appending the TestApplet to the
path to the html file (the layout test file). By
specifying "." as the CODEBASE it is able to find and 
load the class that resides in the same directory and
all is well.

Review URL: http://codereview.chromium.org/48148