Apply ICU patches for ICU tickets 6175 (ISO-2022 and HZ empty segment handling bug) and 5691 (Mulibyte encoding invalid sequence handling bug).
The patch for ICU ticket 5691 was back ported before (r4302,
crbug.com/2074), but that for 6175 has not been. I was mistaken that Webkit-based browsers are not affected by this issue, but recently I found that Apple had
fixed it in their copy of ICU in May 2008 (thanks to Adam). And more importantly, we are affected even though I can't come up with an actual exploit case in a web browser (On the server-side, this can be a real issue) as opposed to a simple alert box coming up.

To make applying the patch easier, I consolidated the patches for 
ICU tickets 5691 and 6175 along with patches for other ICU tickets
listed below into a single patch file (uconv.security.patch). We can
live without the patches for tickets other than 6175 and 5691, but it makes easier to maintain our copy of ICU with them applied together. As a side-benefit(?), GB-HZ converter that was not fixed in r4302 is also fixed (it does not really matter because we don't support GB-HZ, but just in case ...) 

http://icu-project.org/trac/ticket/6001 (r22849, 22851)
http://icu-project.org/trac/ticket/6002 (r22847, 22850)
http://icu-project.org/trac/ticket/6071 (r23041)
http://icu-project.org/trac/ticket/6175 (r23571, 23572)
http://icu-project.org/trac/ticket/6358 (r24245)
http://icu-project.org/trac/ticket/6412 (r24468)
http://icu-project.org/trac/ticket/5691 (r24733,24769,24770,24773,24778,24824,24825)

TEST= 
1. Go to https://bug381412.bugzilla.mozilla.org/attachment.cgi?id=267398
and see a bunch of 'inverse question marks' for U+FFFD. Also, hovering the mouse over 'aaa' should not bring up an alert box. A real layout test will be added later.
2. chrome/fast/encoding/Invalid_multibyte-over-consumption.html 


BUG=2074,3823

Review URL: http://codereview.chromium.org/52030