Fix data url security tests.
The problem was that the security token in V8 was set to the string
"null" for all data urls.  Therefore, the access check succeeded.  The
cause of the regression was that securityOrigin's toString
implementation was changed in WebKit to return "null" instead of
String().

R=kasperl@chromium.org
TBR=scherkus@chromium.org
Review URL: http://codereview.chromium.org/53049