Make the bookmarks bar disappear when the load after the new tab page commits rather than when it is pending. This makes it change at the same time the page changes.
To support this, we now have to keep track of both a pending and a committed
DOMUI object. This is tracked by the RenderManager, which does a similar
swapping between pending and committed RenderViewHosts.

Previous review URL: http://codereview.chromium.org/42512
The only difference is I swapped the order of creating the DOM UI in
RenderViewHostManager::Navigate.

BUG=8963

Review URL: http://codereview.chromium.org/42623