Hook up ctrl+tab and ctrl+shift+tab for switching tabs in linux
We can't just add accelerators for these because GDK_Tab and GDK_ISO_Left_Tab are not valid accelerator keys:
http://www.google.com/codesearch/p?hl=en#ycarM8Ghiog/gtk+-2.4.0/gtk/gtkaccelgroup.c&q=gtk_accel_groups_activate&l=805

Instead, handle these keys in keypress and forward the event on to BrowserWindowGtk::OnAccelerator directly.  We stash a pointer to BrowserWindowGtk on the GtkWindow.

Review URL: http://codereview.chromium.org/42611