Eliminate KeyIdentifer.{h,cpp} in favor of adding an optional method to WebKeyboardEvent.
Comparing the Mac port to the others, I found that there is not perfect
agreement over how to set the keyIdentifier string for a particular key event.
If I changed the Mac port to generate the keyIdentifier string from the
windowsKeyCode, then I would not get the same results as we do now.  Case in
point, the same virtual key is called "Meta" on Mac and "Win" on Windows.
There are also a number of keyIdentifiers on Mac that don't have equivalents on
Windows.

My solution therefore is to continue supporting keyIdentifier on
WebKeyboardEvent but make it easier for tests to synthesize keyIdentifier
strings from the windowsKeyCode.  That matches what we were doing anyways
for tests, so this isn't changing any logic.

BUG=9460
R=avi,dglazkov

Review URL: http://codereview.chromium.org/56051