In certain cases, the coordinates used for pattern rendering can go negative, eg., when a negative translation is applied in the shader matrix.  This causes the rasterizer to blow up, since it accesses memory outside the pattern bitmap.
Since the integer modulus operator for C++ has unspecified behaviour with
negative arguments, its value may go negative.  In this case, we must offset
by the modulus to make it positive again.  I decided to do this at a low level
in skia, since it seems better to make skia robust than to pray that you get
strictly non-negative translations.

(There is a fix for this in src/skia/tile_patch.diff, but it's wrong.)

This patch should be upstreamed to skia.

Review URL: http://codereview.chromium.org/55044