FFmpegDemuxer now uses a thread to handle demuxing.
Since FFmpegDemuxer no longer uses the pipeline thread to demuxer, I had to add a WaitableEvent to TestReader to make sure the read completed (or not!).

Downside is that the test for when a read is never completed now waits 500ms before timing out.  Oh well :(

Review URL: http://codereview.chromium.org/55047