Refactor AppModalDialogQueue and move JS Alert boxes into a MVC.
JavascriptMessageBoxHandler (handles alert, confirm, prompt, and onbeforeunload) was a views class.  This change converts it into an MVC so we can port to linux/mac.

AppModalDialog is the model+controller, JavascriptMessageBoxDialog is the windows specific view.

The onbeforeunload dialog (JavascriptBeforeUnloadHandler) was a subclass of JavascriptMessageBoxHandler that had a different title and button text.  I merged this class into JavascriptMessageBoxHandler by passing a bool to handle the custom button text.

Review URL: http://codereview.chromium.org/63033