Fix the findbar crash.
FindBarController is no longer owned by the BrowserWindow, it is
owned by Browser.  So we can remove the FindBarController code in
BrowserWindowGtk.  Also, go ahead and implement
BrowserWindow::CreateFindBar (which creates the gtk widget and adds
it to the gtk widget hierarchy) and enable the code on linux.

Since the widgets are created and added to the hierarchy later,
we attach the signals after the widgets have been "realized".

Review URL: http://codereview.chromium.org/62126