Relanding the following revisions.
r13555

Log:
Remove the Activex shim registration for generic Activex controls. The Activex shim only handles windows media
player in the wild and will continue to do so for the forseeable future.

This fixes http://code.google.com/p/chromium/issues/detail?id=8584, which is basically related to our NPAPI plugin
installer not getting instantiated on pages which instantiated flash using an object tag, which would result in
webkit attempting to instantiate the activex shim. The shim only loads the windows media player classid, thus causing
this issue.

Added a test shell test which tests whether the default plugin loaded correctly. This is done by attempting to instantiate
a special mime type application/chromium-test-default-plugin, which serves as an indicator to the default plugin to
call a function in the page indicating that the test succeeded. I also made a change to ensure that the default plugin
is loaded in test_shell_tests.

Removed the activexshim dll project from chrome.sln

Bug=8584

New Revision: 13558

Log:
Removing the Activex shim plugin from the list of default plugins caused some activex shim tests to fail. This basically
fails for pages which instantiate the media player like an activex only. To handle this case we attempt to map the clsid
to a NPAPI mime type if possible in the shim. If we succeed then the shim is instantiated as an NPAPI plugin as before.
It internally loads the media player activex.

The other changes are to ensure that the default plugin is only enabled in test_shell for windows for now.

Bug=8584

Review URL: http://codereview.chromium.org/72004