Rebaseline some tests on linux.
inlineBoxWrapperNullCheck.html is a test that changed due to drawRect.

generated.html is the same output as windows, just with anti-aliased
fonts on linux.

The other three (bug56201 bug23847 008) are tests that have checksums
checked in, but no png.  I'm checking in the pngs so they're there
for comparison.

Review URL: http://codereview.chromium.org/75025