Start creating results rows, making sure their visual representation is clipped by the containing popup's roundrect. No text yet, though they seem to actually work when clicked on.
This is not the final construction I'd like, but I'm checkpointing here before the changelist gets any bigger.

In the end, I'm thinking AutocompletePopupView will end up being implemented by the popup's contents view rather than the popup itself. I want to do this because the popup, the contents view and the individual result item views need access to the model and as it stands I need to wire model data methods all the way from the result view up to the popup which is annoying. But for now, I have some of that wiring. I'll do this adjustment after this one, since it requires creating some new files.

Review URL: http://codereview.chromium.org/79023