Pass the ExtensionID to the ExtensionFunctionDispatcher.
This is needed for PageActions that deal with IDs, so
that we can check if one PageAction is trying to change
another PageAction. This converts the object 
ExtensionFunctionDispatcher in RenderViewHost from a
member to a scoped ptr so we can reset it on navigate.

BUG=None
TEST=Extensions should work as before (no visible
change).

Review URL: http://codereview.chromium.org/93125