Merge r13459.
Remove scrollHeight hack from textareas. The hack was trying to
mimic an IE quirk, which it did correctly, but in IE the quirk
is a scrollHeight quirk that is not limited just to textarea
elements. So, it is wrong to mimic it just for textareas and
we can't realistically mimic it for all HTML elements as it would
almost definitely break sites that code for Gecko/WebKit.

The one case that we know of that used to break from not having
this hack was gmail chat boxes growing when they shouldn't.
This bug has been fixed on the gmail side.

TBR= ojan
Review URL: http://codereview.chromium.org/100020