Merge r13644.
Fix a crash where the ResourceMessageFilter is deleted before a
SafeBrowsing check has completed. The problem occurs since the
SafeBrowsingResourceHandler is not deleted when its associated
URLRequest is cleaned up *and* a SafeBrowsing check is in progress.
When the check completes, the next resource handler in the chain
(the AsyncResourceHandler which caches a pointer the now deleted
ResourceMessageFilter) will crash.

This CL adds a notification for objects to know when the
ResourceMessageFilter is destroyed.

BUG=8544 (http://crbug.com)
Review URL: http://codereview.chromium.org/100098