Artificially increase the ref count on WebCore strings that we use as external V8 strings.
We seem to be occasionally losing the data for our external strings.
The current hypothesis is that there is a reference counting bug in
WebCore somewhere which is leading to premature deletion of the string
data.  This change is an attempt to verify that this is in fact the
case.  By artificially increasing the ref count on the strings, we
should reduce the likelihood of accidental deletion because of ref
counting being slightly off.  If we can confirm that this removes most
of the crashes, we know that the problem is WebCore ref counting
related.

BUG=9746
Review URL: http://codereview.chromium.org/99174