Closing the last tab with a download in-progress would cause the tab to be closed and become unusable if the user decides not to proceed with the browser shutdown. This is because we check for in-progress downloads when the browser is closed, and the tab is closed before that, leaving the tab-strip in a bad state.
This CL ensures we also bring-up the confirmation dialog when the last tab is
closed.

BUG=10680
TEST=Start downloading a big file. While the file is downloading, close all
tabs. When closing the last tab, the in-progress download dialog should be
shown. Select the 'Wait for downloads', the download tab should be shown and
the previous tab should still be displayed and functional.
Close all tabs again, this time select the 'Close and cancel downloads'
option, the browser should be closed.


Review URL: http://codereview.chromium.org/99195