Push V8 version 1.2.2 to chromium.
This contains a bug fix for the disappearing external strings bug
(http://crbug.com/9746) as well as a couple of other bug fixes.

In addition to bug fixes, this version of V8 allows the call as
function handler registered on non-function objects to be called in
construct calls.

BUG=9746
Review URL: http://codereview.chromium.org/109011