Fix the "pick a new user data dir" dialog on windows.
There were three things to be fixed:
1) We don't need to run a separate message loop after the
dialog closes.  The MessageLoopForUI handles this properly.
2) When tearing down browser_process_impl, io_loop
hasn't been initialized yet, so we check to make sure it
exists before using it.
3) We don't call window->Close() because that deletes the
dialog class before we're done using it.  This means we leak
the window, but that's ok because we're going to exit anyway.

Review URL: http://codereview.chromium.org/114004