Add baseline for http://trac.webkit.org/changeset/43168
When I run the test in test_shell, I notice that none of the input fields are
marked as misspelled.  I think this is because the spellchecker is not hooked
up to test_shell.  We should probably hook up the spellchecker so tests like
this will be meaningful.

TBR=hbono

Review URL: http://codereview.chromium.org/113090