Merge r15703.
NOTE: filter_false_positive_perftest.cc was copied into the branch
independently. It is not part of this CL, but it is up-to-date.

Decrease the SafeBrowsing false positive rate.

This CL replaces the weak hashing functions of the bloom
filter for a more thorough mixing function. The bloom
filter is modified to use and store random keys in the
mixing function. This addresses the need for randomizing
the contents of each user's bloom filter to prevent a
false positive for a popular site causing a global spike
in hash requests.

BUG=10584 (http://crbug.com/10584)
BUG=10634 (http://crbug.com/10634)
Review URL: http://codereview.chromium.org/115163