Fix regression where exact-match URLs could fail to be promoted when longer URLs were typed more.
We should alwys look up whether we have an exact match, regardless of the input type, and use it if we have it.  My change to not add the "exact match" entry for UNKNOWN input should have only affected the case where that match doesn't, in fact, exist.

BUG=1844093
Review URL: http://codereview.chromium.org/113223