Fix for 11710 on the 172 branch.
NOTE: Instead of MessageBoxFlags::DIALOG_BUTTON_CANCEL, this 
uses the existing DialogDelegate::DIALOGBUTTON_CANCEL.





From jshin:
Change the way button width is calculated for the button. Currently, for the
default button (for which
GetDialogButtonLabel returns an empty string), the width is always set to
kDialogMinButtonWidth. It's sufficiently large for 'Ok', 'Close' and 'Cancel' in
most languages, but not large enough for some
languages. For those languages, this CL sets the button width to
max(preferred_width, kDialogMinButtoWidth),
the former of which is calculated using the translated strings for
Ok/Close/Cancel. 
http://crbug.com/11710

TEST: Start Chrome with '--lang=ml' cmd line flag. Open up the option dialog and
see 'Close' button
at the bottom is large enough for the label ('Close' in Malayalam) and some
space around it. 
Review URL: http://codereview.chromium.org/113257