Apply the fix for bug 11863 from hbono@chromium.org:
This issue is caused by a stupid mistake of mine that I forgot adding a
WM_IME_ENDCOMPOSITION handler to update the find results when a composition is
finished (or canceled). This change adds a handler for WM_IME_ENDCOMPOSITION
messages to update the find results.

BUG=11863 "Korean find-in-page: Backspace all the way, Chrome will highlight all
the characters that contain the first consonant"

TEST=Open a Korean page, type Control+F keys, Switch the input language to
Korean, type a right-alt key, type a 'w' key, type a 'k' key, type a 'a' key,
type a backspace key, and verify there are not any highlited characters. 

Review URL: http://codereview.chromium.org/113498