Fix for breakpad not generating a minidump in certain cases on OSX.
1)Fix logic inversion in IsCrashReporterEnabled().
2)Don't intercept SIGPIPE in non-branded builds since it isn't fatal.
3)Roll DEPS to pickup a bunch of Mac Breakpad fixes.

When breakpad is disabled, we intercept a bunch of signals so that we can crash fast, without waiting for Apple's crash reporter.  The problem was that the function we where using to test whether breakpad was enabled was wrong so we were always installing these signal handlers which where just calling exit().

By fixing the IsCrashReporterEnabled() call, we no longer install these signal handlers if Breakpad is enabled.

In any case SIGPIPE is non-fatal so we remove it from the list of signals we intercept.

There have been a number of fixes to the OSX version of Breakpad recently, so we pull those in as well.

BUG=11929

Review URL: http://codereview.chromium.org/115493