Ensure proper extension when saving an HTML page.
When an HTML page has a title, it is used as the suggested file name in saving
the page. Then the file extension is guessed from the suggested file name.

This results in a wrong extension if the title contains a period ".".
This the source of Bug 10581.

Merely setting ignore_suggested_ext to true in win_util::SaveFileAsWithFilter
is not a good fix, because then we have an issue in saving a page without
a title, where the extension can be correctly derived from the suggested
file name.

This change solves the issue by appending ".htm" to the suggested file name
if the page content type is HTML and the suggested file name doesn't have a
proper extension.

BUG=10581

--------

patch by yuzo@google.com
review here: <http://codereview.chromium.org/115235>