Move ExtraData from WebRequest to WebDataSource.
This adds a DidCreateDataSource method to WebViewDelegate,
which the embedder uses to know if their LoadRequest
resulted in a navigation.  The embedder then associates the
ExtraData with the WebDataSource at that point.  The
WebDataSource then proceeds to be treated as the provisional
data source, possibly failing or being committed.  We then
inspect WebFrame::GetDataSource in DidCommitLoadForFrame to
recover the ExtraData.

We have to take care to handle reference fragment
navigations since those do not result in a new WebDataSource 
being created.  So, in DidChangeLocationWithinPage, we
update the ExtraData associated with the WebDataSource
returned by WebFrame::GetDataSource.

One thing that is important to note:  In
DidCommitLoadForFrame, we would previously always inspect
the value of GetDataSource returned from the main frame
instead of looking at the frame passed to
DidCommitLoadForFrame.  This explains why WebFrameImpl::
LoadRequest needed to cached ExtraData on the current
WebDataSource, and I think doing so is awkward and wrong.

My change is to always store the ExtraData on the first
WebDataSource to be created as a result of LoadRequest. 
Then in DidCommitLoadForFrame, I always inspect the
ExtraData from the given WebFrame.  This means that for
history navigations that only navigate a subframe, we'll end 
up with ExtraData associated with the WebDataSource of a
subframe.  I think this is OK even though the old code was
trying to avoid this scenario.  See the DCHECK removed in
RenderView::UpdateURL.

BUG=11423
R=brettw

Review URL: http://codereview.chromium.org/115575