Merge cl 17052/53 to the 172 branch.
Disk cache: Reset mask_ when the disk cache is being restarted.

It is possible that the size for the index table is reduced when
the cache is restarted (the available disk space could be lower
than what it was when the cache was originally created). If that
is the case, not resetting mask_ to 0 will prevent the mask to
be set correctly for the new table size, resulting in a crash.

BUG=12378
TEST=unittests


Review URL: http://codereview.chromium.org/115939