Fix an issue where themes would sporadically fail to install.
Trying to send decoded images over IPC didn't work too well.  Instead, we'll
write them to a file and have the browser slurp them in from there.  My first
instinct was to use SharedMemory, but that would require us to impose a limit
on the size of the decoded image data.

Also made sure that the undecoded images are deleted when we install.

BUG=13455
TEST=Try the repro steps in bug 13455 several times and make sure it works
every time.
Review URL: http://codereview.chromium.org/119255