PathDifference cuts a supplied "root path" from a second path and returns what remains. It is assumed that this can be used as a relative path from the "root path" to the second path.
If the root path ends with a path separator (\ or /), the function used to cut an additional character from the path and the returned value would be incorrect. This fix makes sure the behavior is consistent and correct no matter if the root path ends with a separator or not.

(This fixes the "epot_tools" tools bug.)
Review URL: http://codereview.chromium.org/118376