Fix regression where hitting enter in the Omnibox would ignore recent editing.
Sadly, the only way I could find to fix this was a latest_result() accessor, because when the user hits enter, we really do need to use the very latest results.  This just highlights how more of this stuff should move to the AutocompleteController.

BUG=13428
TEST=Type "food.c" and wait for the popup to stabilize.  Then _very_ quickly type "om" and hit enter.  You should navigate to food.com instead of searching for "food.c".
Review URL: http://codereview.chromium.org/118398