Copy login-related section of WebDatabase into a new class for use in PasswordStore.
The platform-specific inheritence here is because the current Windows implementation of PasswordStore is a temporary shim, and will eventually need to use its own database instead; since we'll need almost all the fields on the Mac for the bag-on-the-side anyway, it makes sense to share as much of this implementation as possible.

The code and unit tests are pretty much unchanged from WebDatabase (except for the inheritence stuff); the duplication should be temporary as the WebDatabase copy will be removed once Windows switches (and migrates their passwords).

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/119392