Linux: dump_app_syms would fail if the output didn't exist.
This script was trying to only rebuild the symbols when the binary
file was more recent than the symbols file. However, -nt fails when
the right hand side doesn't exist. Therefore it wouldn't actually dump
symbols and the sed script would fail later on.

This adds a test for the non-existance of the output file first.

TBR=mmoss