linux: make extensions service test pass in Debug.
It only passes on the build bots because they still have old directories, I
think?  AbsolutePath() on POSIX will fail if the path doesn't exist.  We let
the program march on in that case and the normal error handling (that would
handle the non-existent path) make the tests continue to pass.

Review URL: http://codereview.chromium.org/132054