Tentative to make the buildbot faster to load.
With the current code, each waterfall/console view page is actually 4 requests to the buildbot
master, and since the buildbot is single threaded, that can slow down a lot the buildbot.

I'm now changing the 3 iframes at the top of the page to point to build.chromium.org instead.

Since build.chromium.org caches the data for 2 minutes before re-asking the buildbot master,
this should decrease the number of requests on our buildbot master by 50-75%. This is huge.

The cons:
- The top bars will update every 2 minutes
- Clicking on a box will load build.chromium.org instead of the local waterfall if you 
are using it internally, so no "force build" from there.

The pros:
- Top bars will load faster
- The main waterfall will receive a lot less traffic, and might be faster.


Review URL: http://codereview.chromium.org/141002