posix: two related changes to make plugin IPC work on POSIX.
[retry, fix windows compile failure]

* use a new ChannelHandle type when passing IPC channels over IPC
The current POSIX code assumes that one end of a channel is always a new
child process (a renderer).  For plugins we need to be able to construct
channels between each of the browser, plugin, and renderer.

This change augments the messages related to creating channels to allow
passing in a base::FileDescriptor containing the socket.  The intent is
that the browser process, as the initial interchange between plugin and
renderer, creates the socketpair() on their behalf and hands each their
respective end of the connection.

* register channel endpoint names in the global pipe map
The plugin code assumes it can map from a string to a channel endpoint
at basically any time.  So whenever we get a channel endpoint over IPC,
we install it in a global map of channel endpoints.

Review URL: http://codereview.chromium.org/113157