Linux: add linux-gate.so to the minidump mappings.
Add linux-gate.so to the minidump mappings list by checking for the
load address in the aux vector and adding a filename of linux-gate.so
when we encounter that load address. This is necessary for crash
reporting to walk the stack correctly when a crash happens while in a
system call.

http://codereview.chromium.org/147032

Committed for Neal Sidhwaney.