Use regular notifications for theme changes.
This should have no impact on the current behavior.

On linux, we'll want to get these notifications in a few places other than
just the root browser windows so having it as a regular notification
makes it easier for us to just listen for the event.

Review URL: http://codereview.chromium.org/149032