linux: fix renderer/plugin crash when loading multiple pages with plugins
Currently we crash if one renderer has a plugin and another one has a plugin that gets destroyed, and started again (e.g. reload). We get away with reload on a single page currently because the process gets torn down.
This properly cleans up the socket and channel mapping on both sides, so it can be restarted.

Review URL: http://codereview.chromium.org/155174