plugins: use OnChannelError to detect when the channel goes away
Previously we used a special Watcher object to watch the process,
but that is not portable to POSIX and this is simpler anyway.
With this change, I now see ~PluginProcessHost() running when a
plugin crashes.

If you recall all the way back to
  http://codereview.chromium.org/16814
we did a similar thing to the renderer host.

Review URL: http://codereview.chromium.org/155331