test_shell: set a title on the home page
This seems pointless, but for debugging purposes internally we
use the page title as an annotation on OS-level windows which helps
track them down when looking at a list of all windows.  For
reasons too complicated for me to care about this doesn't work when
a page lacks a title, hence this fix.

Review URL: http://codereview.chromium.org/149642