linux: destroy browser-side plugin windows on plugin crash
Because the plugin windows must be destroyed by the UI thread,
we store the map of plugin PID -> plugin windows there,
on the RenderWidgetHostView.

This differs slightly from Windows, where the map is stored
on the PluginProcessHost, but is otherwise pretty similar.

TEST=crash a windowed plugin; you should see a sad plugin
icon in the place of the plugin

Review URL: http://codereview.chromium.org/155517