Modify change in CL 21434 - allow standard font smoothing as well as none
Prior to CL 21434 test_shell did not actually check if ClearType was
disabled. I had thought that the requirement for the layout tests was
that *all* font smoothing was disabled (both ClearType and standard). Turns
out that standard is okay, so I am reenabling it here.

Note that CL 21434 was really just a re-reverting of 21368.

TEST=none
BUG=none
R=tony@chromium.org

Review URL: http://codereview.chromium.org/160082