Merge 21261 - mp4 types in omnibox only for chrome, not chromium
The types are still claimed as supported because layout tests require them, but the extensions will not map to mimetypes, so typing them into the omnibox will download instead of trying to play the video.
The fixes for canPlayType that kyle is doing will fix some of this, and we'll also need to update the media in the layouttests.  So that portion of mimetype will be deferred

BUG=17323
TEST=in the omni box type in the name of an mp4 and it should work in chrome, but not in chromium.

Review URL: http://codereview.chromium.org/155835

TBR=fbarchard@chromium.org

Review URL: http://codereview.chromium.org/159410