Quick work around for av_seek_frame() causing video to go completely out of sync.
If we have a freshly opened AVFormatContext, telling it to seek to the beginning (pts=0) apparently causes the video and/or audio streams to get timestampped incorrectly, causing video sync issues.  If the video has been playing a bit, a seek to the beginning causes no issues.  Go figure.

BUG=17832
TEST=ogg videos should remain in sync

Review URL: http://codereview.chromium.org/159438