Merge our C+J segmentation changes to ICU 4.2.1.
This will be followed by other merges. When everything (quite many patches we applied to 3.8 are now included in ICU 4.2.1. So, there are fewer patches to apply to 4.2.1 than to 3.8.x) is merged back, we'll change DEPS and build files to use ICU 4.2.1 instead of ICU 3.8.x

BUG=8198
TEST=Build ICU and run tests on Linux and all tests pass.
       $ cd $ICUROOT/source
       $ ln -s ../public/common/unicode ./common/
       $ ln -s ../public/i18n/unicode ./i18n/
       $ ln -s data/brkitr/cjdict.txt test/testdata/cjdict-truncated.txt
       $ cd <Your build directory> # to avoid cluttering your source tree.
       $ $ICUROOT/source/runConfigureICU Linux
       $ make checks

Review URL: http://codereview.chromium.org/159567