First stab at a layout tests flakiness/speed dashboard. This isn't functional yet, but I want to get this reviewed and in the tree so I can do the rest incrementally.
This works by having the bots generate JSON that is
then red into a static HTML file that generates
the dashboard from the JSON.

I've tried to make this generic, so we should be able
to use the same HTML file for our other test types
(e.g. UI tests) as well once this is functional by
just having the bots that run those tests generate
the JSON files and copy them to the right place.

All the work that needs doing to get this 100% functional
is listed as a TODO at the top of flakiness_dashboard.html.

Most of what's left is buildbot integration (i.e. copy
files to the right place on the bot). 
Review URL: http://codereview.chromium.org/149656