Ensure that we never call into WebCore::Page static methods when Page constructor hasn't been called.
This is a top crash for Mac right now. The crash occurs because the static table of all pages in the renderer is initialized 
in WebCore:: Page constructor, and if we attempt to access this table before the Page was created (i.e. before WebView::Create is called),
we get into a crashy situation. I am really not sure how to reproduce this, but I can defend against the situation.

BUG=17555
R=darin
TEST=unable to repro, speculative fix.

Review URL: http://codereview.chromium.org/159887