More Linux theme fixes.
(1) If the theme provides a inactive tab background, don't offset the
image.  This matches the Windows code.

(2) Tile the toolbar background in both the X and Y direction (no longer
use a ninebox for this).

(3) Make the toolbar event box have an invisible window so toolbar
backgrounds with transparency will show the background behind it.

TEST=Install the Grass theme from the theme gallery.  It shouldn't look janky.

Review URL: http://codereview.chromium.org/164023