The UrlRequestAutomationJob::GetResponseInfo function needs to return a valid X509Certificate in the HttpResponseInfo structure passed in if the current URL scheme is https
Not returning this certificate causes a number of DCHECKS to fire in the network stack implementation and in the renderer process, which eventually causes the navigation to fail.

Fix is to return a dummy X509Certificate on the same lines as Gears.

This fixes bug http://b/issue?id=2039699

Bug=2039699

Review URL: http://codereview.chromium.org/164110