Merge 22671 - Fix a leak in AudioRendererBaseTest::OneCompleteReadCycle
We should a local reference of the input buffer, so that buffer
gets cleanup after it is out of scope.

Also disabled AudioUtilTest.FoldChannels_s16_benchmark because
it is running too long. It shold be run locally instead of on
the build bots.

TEST=media_unittests gtest_filter=AudioRendererBaseTest.OneCompleteReadCycle
TBR=scherkus

Review URL: http://codereview.chromium.org/165087

TBR=hclam@chromium.org

Review URL: http://codereview.chromium.org/165425