Argh, reverting the change to remove the "crx" file extension hack in DownloadManager.
For those playing along at home, this change makes it so that
we once again consider any file that ends in ".crx" an
extension.

The problem is that file:// URLs are not-sniffable, so with
this change we cannot install extensions from file:// URLs.

We need to fix things so that we consider file associations
for local files, I guess.

TBR=mpcomplete@chromium.org