Fix build bustage.
r22016 forces sizing of a new tab as soon as TabContents is created.

On the 172 branch, we need to get the view() from a WebContents.

I also null-check AsWebContents() before use.

Ben or Brett: I need a review of lines 1389-1422 to be sure that
this will work as expected under the old code.

For reference, here is Nate's original code review: 
http://codereview.chromium.org/126290


Review URL: http://codereview.chromium.org/165483