Stop resource loading for WebCore::MediaDocument
BUG=17973

When a WebCore::Document is loaded, a resource request
for the document is made. In case of MediaDocument the
request for the media file is made separately through
<video> and so the DocumentLoader should be stopped.
This change mimic what is done in WebHTMLRepresentation
(which is similar to our RenderView) and stops the
resource loading in WebFrameLoaderClient since RenderView
doesn't have access to WebCore types.

Review URL: http://codereview.chromium.org/165286