Check for NULL Profile::GetDefaultRequestContext().
This papers over a chrome-bot crash.

The problem is that the ChromeURLRequestContexts have a split life between the UI thread and IO thread.
Although conceptually they live on the IO thread, they are lazy-initialized on the UI thread, and are contained by a Profile which also lives on the UI thread.

What happens in this crash, is ResolveProxyMsgHelper (the class which handles IPCs from the plugin process for proxy resolving) calls Profile::GetDefaultRequestContext() to get at the main URLRequestContext from the IO thread. Well, during the shutdown sequence, ~ProfileImpl NULLs the default request context before the IO thread is torn down, so consumers on the IO thread may get NULL. Or even worse, might get a non-NULL pointer to a request context that has already been freed.

With this patch I hack past the NULL case.

Really we need a proper solution to managing URLRequestContexts split personality... I did a quick survey of the code and found other consumers that use GetDefaultRequestContext from the IO thread, so this bug may be happening elsewhere.

BUG=http://crbug.com/18358

Review URL: http://codereview.chromium.org/172085