Fix a bug where we reused a va_list across multiple calls.
The problem is that va_arg modifies the va_list so that the next call to va_arg
returns the next argument.  It looks like this code had attempted to not reuse
the list by making a copy, but it passed the original instead of the copy.

Review URL: http://codereview.chromium.org/174342