Fix "crashed extension" infobar browser crashes.
This is a general rework of how "crashed extension" infobar works
and how the extension is actually recovered after the crash.

This is a subset of original http://codereview.chromium.org/164151/ .
I had to remove the part which unloads the entire extension on crash,
because it interacts badly with other parts of the browser. I'm fixing that.

TEST=See bug.
http://crbug.com/15888

Review URL: http://codereview.chromium.org/173314