Add isVisible parameter to WebPlugin::updateGeometry.
This allows for geometry updates while the plugin is hidden so that when it
becomes visible, the plugin is properly positioned.

In this change, I also revised RenderWidget::SchedulePluginMove to properly
merge a move that only updates the visible state, and I fixed up TestShell
to properly handle moves that only update the visible state.

R=jam
BUG=19888,20260
TEST=covered by layout tests

Review URL: http://codereview.chromium.org/174513