Fix crash-on-launch on Snow Leopard
The signal-self syntax needed to fix bug 20370 doesn't work on Leopard and
causes the Sandbox to barf so renderers don't star tup.

We introduce a ;10.5_ONLY keyword to the sandbox that is substituted at runtime (the line with the rule is effectively uncommented).  This allows us to use leopard or snow-leapard-only sandbox syntax

BUG=20516
TEST=Chrome should no longer crash on launch under SL.

Review URL: http://codereview.chromium.org/183003