[Mac] Handle shift-delete when omnibox popup is up.
When the user hits shift-delete (that's forward delete),
attempt to delete the item highlighted from the list.  This only seems
to work for history items.

http://crbug.com/20281
TEST=No idea.  I've managed to delete an item, but most items seem
immune to deletion, this code is calling the right place, but the
match itself checks !deletable.  Sometimes I have luck going to a
site, then typing part of the URL and using shift-del on it.

Review URL: http://codereview.chromium.org/176022