Merge 24568 - Make audio and video in sync while playback rate != 1.0
BUG=20290
TEST=play a video with rate 0.5, audio and video should be in sync

This change use the playback rate as a scaling factor for
delay introduced in the audio hardware buffer. After we
scaled the amount of audio buffer not played, we'll be
able to make audio / video in sync again.

Review URL: http://codereview.chromium.org/174584

TBR=hclam@chromium.org

Review URL: http://codereview.chromium.org/176049