Detect perf regressions and speedups automatically.
Compare the results we get with the results present
in perf_expectations.json.  Update the summary box
according to which result we encounter.

BUG=18597
TEST=Test these changes using a set of tuned perf
expectations files: one null entry, the expected
range covering the perf result, the range below the
perf result, and the range above the perf result.
In the first two cases, no visible changes appear in
the summary box.  In the third case, the summary box
is red and regressions are listed.  In the fourth case,
the summary box is orange and speedups are listed.
Review URL: http://codereview.chromium.org/177028