When forking, restore the default exception handler port.
The browser process launches children with fork && exec.  On the Mac, this
results in a child that has its exception ports set the same as the parent.
When Breakpad is enabled, this means that a child such as a renderer process
begins its life with its exception port pointing to a Breakpad instance
running inside the browser.  This is not intended.  Even if the child installs
its own Breakpad handler (which a renderer would), the previous handler would
be pointing to the instance inside the browser, and if the child's Breakpad
is misconfigured to pass exceptions to the previous handler it will wind up
being passed back to the parent.  The misconfiguration was corrected in r25311.
This change restores the default exception handler port, which should be
Apple Crash Reporter/ReportCrash, in the child process immediately after it
it has been forked.

BUG=20371
TEST=None, because r25311 fixed the visible aspect of this bug
Review URL: http://codereview.chromium.org/200007