Hook up WebEditingClient.
Moves the WebViewDelegate parameter back to WebView::Create and
adds a second parameter for WebEditingClient.

I had hoped to make the WebEditingClient NULL for RenderView on
Windows and Mac, but that turned out to not be an option.  I need
a few methods on all platforms.

The Describe* functions from EditorClientImpl are moved into the
TestWebViewDelegate since they are only applicable to layout tests.

R=dglazkov
BUG=
TEST=none

Review URL: http://codereview.chromium.org/195008