Size the buffer used for the metrics log text properly.
The existing code was abusing the WriteInto interface (base/string_util.h).
The size passed to WriteInto should be the size of the string plus one for
a terminating NUL byte.

This bug caused the final newline in the log to be clipped.  c_str users
(there probably weren't any) would see the final newline which wouldn't be
wouldn't be followed by a NUL byte.

TEST=not really any
BUG=21733
Review URL: http://codereview.chromium.org/194099