Revert 26295 - Build sync engine as part of the browser build.
As before, syncapi is built as a dynamic library, and only
on windows.  The difference is that it's built from source,
rather than being pulled in as a binary.

Changes to sync engine code:
 * Use OS_WIN instead of OS_WINDOWS (requires build_config.h).

 * Rename platformspecific files to match the chrome convention.
   This allows them to be excluded by rules that already exist
   in chrome.gyp.  The convention is either a /win/ directory,
   or an _win.cc at the end of the file.  Other valid platforms
   are _mac, _linux, and _posix

Changes to DEPS:
 * On Windows, pull in pthreadswin32.

Changes to chrome.gyp:
 * Add new library targets for sync, notifier, sync_proto, and
   syncapi.

Review URL: http://codereview.chromium.org/193103

TBR=nick@chromium.org

Review URL: http://codereview.chromium.org/212025