Fix the docs.  Apparently they've been broken since August 3rd rev 22348.
The scary thing is that's a little hard to believe
it's been broken that long and no one noticed
so I hope this is the correct fix but it does
appear to be wrong. The '/'s that were removed
are used in a string concat in the HTML templates.

Sometimes they need to be '' and sometimes 'folder/'
so that urls become

http://foo/bar.html

or

http://foo/folder/bar.html



Review URL: http://codereview.chromium.org/213033