[Relanding erg's change with fix for toolkit_views shortcuts and interactive ui tests. Note that this was originally reviewed in http://codereview.chromium.org/217022/show I originally Elliot suggestion of replacing the usage of int for keycode with the bae::Keycode type, but that led to the CL getting out of hands (as this is used in many different places). So this is only the patch set 1 of that CL, I'll replace the type in another CL]
Use windows keycodes under linux (and all non-windows platforms). 

This fixes any place where we use a VKEY_* (RenderWidgetHost, for example) 
under Linux, but breaks accelerators in TOOLKIT_VIEWS which relied on this 
wrong behaviour. 

Previously, keyboard_codes_linux defined all the VKEY_* constants as their 
GDK_* counterparts, which is wrong since the VKEY_* are supposed to resolve 
to windows key codes. 

BUG=22551
TEST=Make sure accelerators still work as expected on Chrome Linux and 
Chrome Linux with toolkit views. Test when the the accelerators with 
the focus in the location bar and also with the focus on the page.
Review URL: http://codereview.chromium.org/235025