Fix --new-baseline to honor the --platform flag.
Also test_types/text_diff.py attempts to be smarter than the user by checking to see if the text results are a render tree output and, if not, writing them into the generic directories rather than the specified directory. Except, that code looks broken, too. I think it's probably best to just do what the user told it to do, so I've removed that path.

Thanks to senorblanco for finding the bug.

  R=ojan@chromium.org
  TEST=none
  BUG=none

Review URL: http://codereview.chromium.org/255068