Reland: HTML Pack Extension Dialog / Linux & Mac Packaging Support.
original issue: http://codereview.chromium.org/207062

The issue had to do with a symbol collison with the nss libraries (which are currently out-of-date) on the build bots.

HTML Pack Extension Dialog. This removes the views implementation of the ExtensionPackDialog, and implements the dialog in html in the extensions_ui DOMUI.

Additionally, support is added for packaging extensions via ---pack-extension on linux and mac

BUG=20668, 20669

TBR=aa,wtc

Review URL: http://codereview.chromium.org/265032