Add a visualizer of the tree open/closure events.
This is exposed at URL:
  /status_viewer.html

The webpage does an XMLHTTPRequest to /allstatus?startDate=XXX&endDate=YYY to get the data that it needs, and formats all the visualizations client-side using javascript.

TODO(eroman): Support has not been added to /allstatus for "startDate=XXX" or "startDate=YYY" parameters yet. I plan to do that in another change. What this means until then is the dashboard will still work, but it is going to fetch way more data then it needs; so viewing small time ranges is going to be unnecessarily slow.

BTW: It has been a while since I wrote javascript, so the style probably isn't up to regulation... Hopefully we can be a bit lenient on this front :)

BUG=21797

Review URL: http://codereview.chromium.org/278010