Use a multiset rather than a set to store bookmarks when running the model associator merge.
For cases where there's an exact duplicate URL and title,
this allows each BookmarkNode a chance to be associated
with a sync node.  With a set, we'd forget about all
but one of the duplicates, resulting in the creation
of yet more duplicates.  The result was exponential
growth with each merge and sync.

BUG=24995
TEST=

Review URL: http://codereview.chromium.org/288001