Prevent thread creation if Join has already been called.
Tracking the thread id in the constructor, destructor and Start()
shows that the thread object can be started and stopped on
different threads.  This means that there is a race condition
if Stop() or Join() gets called during Start(), which could cause
the thread to start with a pointer to a deleted object.

BUG=23251
TEST=none

Review URL: http://codereview.chromium.org/274079