Allow slightly larger browser and page action icons.
This allows 19px icons to be able to be used for both
browser and page icons.

I think it looks nicer with the page actions slightly smaller, and that is what we also usually do in Chrome,
but some Chrome location bar icons use 18px for soft edges,
so I guess this will just have to be something we advise
developers on.

We can actually fit up to 21 (whoa nelly) pixels on Windows,
but apparently the space is slightly smaller on mac.

Also minor layout fix. We were sizing the browser action
buttons 1px too short.

BUG=24881
TEST=Load chrome/test/data/extensions/samples/icon_size_test. Icons should be 17px for the page action and 19px for the browser action and centered nicely in the space.

Review URL: http://codereview.chromium.org/286001