Remove special code fro calculating bucket index in linear histograms
IMO, the original code was a premature optimization, and worse yet, it
had a subtle off by 1 bug due to floating point rounding (which was
identified by ERoman.... Thanks!!).

This CL removes the special code from linear histograms, and relies on
the "standard" binary search approach to locate the correct bucket
when a new sample is processed.

r=eroman
BUG=24160

Review URL: http://codereview.chromium.org/273065