Fix exception on missing perf expectations.
Don't crash when perf expectations is missing.  Show a
warning for the build step if the file cannot be
opened or read.  Allow setting the warning for steps
that should be able to read a perf expectations file.
Also fix docs for file format.

BUG=none
TEST=missing perf expectations does not cause a crash
Review URL: http://codereview.chromium.org/285009