Add am/sw locale data and invuca table
1. Add the full locale data (yet incomplete ) for Amharic and Swahili to our ICU.
We're expecting to get a better locale data soon, but we want to get this in for now. 

2. Add invuca table which became necessary due to the patch for 
  http://bugs.webkit.org/show_bug.cgi?id=30437

Windows dll (pre-built) will be checked in in next CL. The 'assembly' source files for Linux/Mac are included but not shown here because they're too large to be shown here. 

BUG=25451
TBR=mal
TEST=1. With '--lang=am' or '--lang=sw' on Windows or 'LANGUAGES=am chrome' on Linux, language names in Options | Advanced | Languages and Fonts menu are actually translated into Amharic and Swahili.
2. After this change is pulled in, LayoutTests/fast/text/international/japanese-kana-letters.html passes on Linux/Mac. 

Review URL: http://codereview.chromium.org/308011